backend_api/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
httpx
python-multipart
huggingface_hub
